+++ src/compliance_tests/TestIso_7_2_2.cpp
@@ -361,7 +361,7 @@ class TestIso_7_2_2 : public test::TestBase
 
             mon_bit_frm->ConvRXFrame();
 
-            int num_stuff_bits = drv_bit_frm->GetNumStuffBits(StuffKind::Normal);
+            size_t num_stuff_bits = drv_bit_frm->GetNumStuffBits(StuffKind::Normal);
 
             /* In FD enabled variant, if last bit of data field is stuff bit, but model has this bit
              * as fixed stuff bit before Stuff count. So count in also each fixed stuff bit even
@@ -371,7 +371,7 @@ class TestIso_7_2_2 : public test::TestBase
             if (test_variant == TestVariant::CanFdEna)
             {
                 Bit *bit = drv_bit_frm->GetBitOf(0, BitKind::StuffCnt);
-                int index = drv_bit_frm->GetBitIndex(bit);
+                size_t index = drv_bit_frm->GetBitIndex(bit);
                 BitVal value = drv_bit_frm->GetBit(index - 1)->val_;
                 if ((value == drv_bit_frm->GetBit(index - 2)->val_) &&
                     (value == drv_bit_frm->GetBit(index - 3)->val_) &&
@@ -383,10 +383,10 @@ class TestIso_7_2_2 : public test::TestBase
             /**************************************************************************************
              * Execute test
              *************************************************************************************/
-            for (int stuff_bit = 0; stuff_bit < num_stuff_bits; stuff_bit++)
+            for (size_t stuff_bit = 0; stuff_bit < num_stuff_bits; stuff_bit++)
             {
-                TestMessage("Testing stuff bit nr: %d", stuff_bit);
-                TestMessage("Total stuff bits in variant so far: %d", stuff_bits_in_variant);
+                TestMessage("Testing stuff bit nr: %zu", stuff_bit);
+                TestMessage("Total stuff bits in variant so far: %zu", stuff_bits_in_variant);
                 stuff_bits_in_variant++;
 
                 /*
@@ -397,7 +397,7 @@ class TestIso_7_2_2 : public test::TestBase
                 mon_bit_frm_2 = std::make_unique<BitFrame>(*mon_bit_frm);
 
                 Bit *stuff_bit_to_flip = drv_bit_frm_2->GetStuffBit(stuff_bit);
-                int bit_index = drv_bit_frm_2->GetBitIndex(stuff_bit_to_flip);
+                size_t bit_index = drv_bit_frm_2->GetBitIndex(stuff_bit_to_flip);
                 /* Here we only flip, no compensation! This is because since we
                  * flip stuff bit, we remove the synchronization edge, therefore there
                  * is no need to compensate the edge position!
