+++ src/compliance_tests/TestIso_7_2_4_1.cpp
@@ -86,7 +86,7 @@ class TestIso_7_2_4_1 : public test::TestBase
         void ConfigureTest()
         {
             FillTestVariants(VariantMatchType::CanFdEnaOnly);
-            for (int i = 0; i < 8; i++)
+            for (size_t i = 0; i < 8; i++)
                 AddElemTest(TestVariant::CanFdEna, ElemTest(i + 1));
 
             CanAgentConfigureTxToRxFeedback(true);
@@ -168,7 +168,7 @@ class TestIso_7_2_4_1 : public test::TestBase
              *************************************************************************************/
             mon_bit_frm->ConvRXFrame();
 
-            int num_stuff_bits = drv_bit_frm->GetNumStuffBits(StuffKind::Normal);
+            size_t num_stuff_bits = drv_bit_frm->GetNumStuffBits(StuffKind::Normal);
 
             /* In FD enabled variant, if last bit of data field is stuff bit, but model has this bit
              * as fixed stuff bit before Stuff count. So count in also each fixed stuff bit even
@@ -176,7 +176,7 @@ class TestIso_7_2_4_1 : public test::TestBase
              * FD enabled variant will be higher than in ISO 16845, but this does not mind!
              */
             Bit *bit = drv_bit_frm->GetBitOf(0, BitKind::StuffCnt);
-            int index = drv_bit_frm->GetBitIndex(bit);
+            size_t index = drv_bit_frm->GetBitIndex(bit);
             BitVal value = drv_bit_frm->GetBit(index - 1)->val_;
             if ((value == drv_bit_frm->GetBit(index - 2)->val_) &&
                 (value == drv_bit_frm->GetBit(index - 3)->val_) &&
@@ -187,7 +187,7 @@ class TestIso_7_2_4_1 : public test::TestBase
             /**************************************************************************************
              * Execute test
              *************************************************************************************/
-            for (int stuff_bit = 0; stuff_bit < num_stuff_bits; stuff_bit++)
+            for (size_t stuff_bit = 0; stuff_bit < num_stuff_bits; stuff_bit++)
             {
                 TestMessage("Testing stuff bit nr: %d", stuff_bit);
                 stuff_bits_in_variant++;
@@ -200,7 +200,7 @@ class TestIso_7_2_4_1 : public test::TestBase
                 mon_bit_frm_2 = std::make_unique<BitFrame>(*mon_bit_frm);
 
                 Bit *stuff_bit_to_flip = drv_bit_frm_2->GetStuffBit(stuff_bit);
-                int bit_index = drv_bit_frm_2->GetBitIndex(stuff_bit_to_flip);
+                size_t bit_index = drv_bit_frm_2->GetBitIndex(stuff_bit_to_flip);
                 stuff_bit_to_flip->FlipVal();
 
                 drv_bit_frm_2->InsertPasErrFrm(bit_index + 1);
